Project working with AngularJS and property with type as a parameter in Property.

+++ src/main/java/main/service/PropertiesService.java
@@ -9,7 +9,7 @@ public interface PropertiesService {
 
     void save(Property property);
 
-    Optional<Property> findById(int id);
+    Optional<Property> findById(Long id);
 
     Optional<Property> findByPropName(String propName);
 }
